Rename pageTypes -> modelTypes
