+++ src/components/Sidebar/menu/hooks/useMenuStructure.tsx
@@ -31,7 +31,7 @@ import { commonMessages, sectionNames } from "@dashboard/intl";
 import { menuListUrl } from "@dashboard/navigation/urls";
 import { orderDraftListUrl, orderListUrl } from "@dashboard/orders/urls";
 import { pageListPath } from "@dashboard/modeling/urls";
-import { pageTypeListUrl } from "@dashboard/pageTypes/urls";
+import { pageTypeListUrl } from "@dashboard/modelTypes/urls";
 import { productListUrl } from "@dashboard/products/urls";
 import { languageListUrl } from "@dashboard/translations/urls";
 import { Box } from "@saleor/macaw-ui-next";
