+++ src/index.tsx
@@ -67,8 +67,8 @@ import { NotFound } from "./NotFound";
 import OrdersSection from "./orders";
 import PageSection from "./modeling";
 import { pageListPath } from "@dashboard/modeling/urls";
-import PageTypesSection from "./pageTypes";
-import { pageTypeListPath } from "./pageTypes/urls";
+import PageTypesSection from "./modelTypes";
+import { pageTypeListPath } from "@dashboard/modelTypes/urls";
 import PermissionGroupSection from "./permissionGroups";
 import PluginsSection from "./plugins";
 import ProductSection from "./products";
