+++ src/modelTypes/components/PageTypeBulkDeleteDialog/index.ts
