+++ src/modelTypes/components/PageTypeCreatePage/PageTypeCreatePage.tsx
@@ -7,7 +7,7 @@ import { Metadata, MetadataFormData } from "@dashboard/components/Metadata";
 import { Savebar } from "@dashboard/components/Savebar";
 import { PageErrorFragment } from "@dashboard/graphql";
 import useNavigator from "@dashboard/hooks/useNavigator";
-import { pageTypeListUrl } from "@dashboard/pageTypes/urls";
+import { pageTypeListUrl } from "@dashboard/modelTypes/urls";
 import useMetadataChangeTrigger from "@dashboard/utils/metadata/useMetadataChangeTrigger";
 import React from "react";
 import { useIntl } from "react-intl";
