+++ src/modelTypes/components/PageTypeDetails/PageTypeDetails.tsx
