+++ src/modelTypes/components/PageTypeList/PageTypeList.tsx
@@ -7,7 +7,7 @@ import { TablePaginationWithContext } from "@dashboard/components/TablePaginatio
 import TableRowLink from "@dashboard/components/TableRowLink";
 import { PageTypeFragment } from "@dashboard/graphql";
 import { getPrevLocationState } from "@dashboard/hooks/useBackLinkWithState";
-import { PageTypeListUrlSortField, pageTypeUrl } from "@dashboard/pageTypes/urls";
+import { PageTypeListUrlSortField, pageTypeUrl } from "@dashboard/modelTypes/urls";
 import { getArrowDirection } from "@dashboard/utils/sort";
 import { TableBody, TableCell, TableFooter } from "@material-ui/core";
 import { makeStyles } from "@saleor/macaw-ui";
