+++ src/modelTypes/hooks/usePageTypeDelete/usePageTypeDelete.tsx
@@ -1,7 +1,7 @@
 // @ts-strict-ignore
 import { PageCountQueryVariables, usePageCountQuery } from "@dashboard/graphql";
 import { pageListUrl } from "@dashboard/modeling/urls";
-import { PageTypeListUrlQueryParams, PageTypeUrlQueryParams } from "@dashboard/pageTypes/urls";
+import { PageTypeListUrlQueryParams, PageTypeUrlQueryParams } from "@dashboard/modelTypes/urls";
 import React from "react";
 
 import * as messages from "./messages";
