+++ src/modelTypes/views/PageTypeDetails.tsx
